school_helper/requirements.txt
aiogram>=3.25.0
openai>=1.50.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
